proxy generator refactor
