+++ OwlTree.Generator/GeneratorHelpers.cs
@@ -22,6 +22,7 @@ namespace OwlTree.Generator
         public const string Tk_OwlTree = "OwlTree";
         public const string Tk_NetworkObject = "NetworkObject";
         public const string Tk_ProxySuffix = "Proxy";
+        public const string Tk_CsFile = ".g.cs";
         public const string Tk_IEncodable = "IEncodable";
         public const string Tk_IVariable = "IVariable";
